+++ tests/testthat/test-report_intercept.R
@@ -49,4 +49,12 @@ test_that("reflevel", {
     as.character(report_intercept(m3)),
     "The model's intercept, corresponding to f = 1, is at 0.17 (95% CI [-0.47, 0.81], t(27) = 0.55, p = 0.584)."
   )
+
+  # Logical predictor
+  mtcars$more_than_4_cyl <- as.logical(mtcars$cyl > 4)
+  m4 <- lm(mpg ~ more_than_4_cyl, data = mtcars)
+  expect_identical(
+    as.character(report_intercept(m4)),
+    "The model's intercept, corresponding to more_than_4_cyl = FALSE, is at 26.66 (95% CI [24.41, 28.92], t(30) = 24.16, p < .001)."
+  )  
 })