+++ NEWS.md
@@ -2,6 +2,8 @@
 
 Bug fixes
 
+* Fixed an issue in `report()` where the reference level for logical predictors was incorrectly displayed as `[?]` instead of `FALSE` for the intercept (@M-Colley, #598).
+
 * Fix `report()` crash when character vector has only one unique value (#578).
 
 # report 0.6.3